Huge commit: Commandline Arguments, Error Checking, and Diagnostics
Added commandline arguments, as well a default configuration if the user
does not add any. Try running with none to see the usage message.

Added better error checking to a number of things that are prone to
throw exceptions, as well as a safe way of accessing sockets where it
sequentially tries to increase the socket until it finds one not in use

Commented out extraneous socket initialization in VisionServerThread
that was causing errors. Comment it back in if this was a mistake.

Added more and more detailed diagnostic messages to help debug